primitive update of pc_news, added additional checks of data/LOCK files that could be added by hand for special purposes (e.g. if somebody runs in somebody elses data disk).

+++ bin/start.csh
@@ -44,7 +44,8 @@ if (! -d "$datadir") then
   echo ">>  but that will most likely end up on your NFS file system and be"
   echo ">>  slow"
   echo
-  rm -f LOCK
+  if (-e "LOCK") rm -f LOCK
+  if (-e "data/LOCK") rm -f data/LOCK
   exit 0
 endif
 
@@ -156,6 +157,7 @@ endif
 
 # remove LOCK file
 if (-e "LOCK") rm -f LOCK
+if (-e "data/LOCK") rm -f data/LOCK
 
 # Detect error status flagged by code (for cases where this does not get
 # propagated to the mpirun status):
